% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_compare_marginal}
\alias{plot_compare_marginal}
\title{.......}
\usage{
plot_compare_marginal(
  plots = NULL,
  abc_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  statistics_labels = NULL,
  xlimit = NULL,
  sample_num = NULL,
  plot_statistics = FALSE,
  plot_hist = FALSE,
  plot_hist_point = FALSE,
  alpha = 0.3,
  plot_prior = FALSE
)
}
\description{
.......
}
