% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_smcrf_joint}
\alias{plot_smcrf_joint}
\title{.......}
\usage{
plot_smcrf_joint(
  smcrf_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  lims = NULL,
  nBins = 5
)
}
\description{
.......
}
