% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_smcrf_marginal}
\alias{plot_smcrf_marginal}
\title{.......}
\usage{
plot_smcrf_marginal(
  smcrf_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  statistics_labels = NULL,
  plot_statistics = FALSE,
  xlimit = NULL,
  alpha = 0.3,
  plot_hist = FALSE
)
}
\description{
.......
}
