% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcrf.r
\name{smcrf}
\alias{smcrf}
\title{Approximate Bayesian Computation sequential Monte Carlo via random forests
for marginal or joint posterior distributions.}
\usage{
smcrf(
  method = "smcrf-single-param",
  statistics_target = NULL,
  smcrf_existed_results = NULL,
  model,
  perturb,
  range = NULL,
  parameters_initial = NULL,
  nParticles,
  parallel,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{method}{Random forest to implement in each iteration.
Option 1 is "smcrf-single-param", which implements ABC-RF for each parameter and results in marginal posterior distributions.
Option 2 is "smcrf-multi-param", which implements ABC-DRF for all parameters and results in a joint posterior distribution.}

\item{statistics_target}{.}

\item{smcrf_existed_results}{.}

\item{model}{.}

\item{perturb}{.}

\item{range}{.}

\item{parameters_initial}{.}

\item{nParticles}{.}

\item{parallel}{.}

\item{n_cores}{.}

\item{...}{.}
}
\value{
Something.
}
\description{
Approximate Bayesian Computation sequential Monte Carlo via random forests
for marginal or joint posterior distributions.
}
