% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcrf.r
\name{smcrf_multi_param}
\alias{smcrf_multi_param}
\title{.......}
\usage{
smcrf_multi_param(
  statistics_target = NULL,
  model,
  perturb,
  range = NULL,
  parameters_initial = NULL,
  nParticles,
  parallel,
  n_cores = NULL,
  save_model = TRUE,
  save_rda = FALSE,
  filename_rda = "ABCSMCDRF.rda",
  splitting.rule = "CART",
  smcrf_multi_param_results = NULL,
  ...
)
}
\description{
.......
}
