% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_compare_joint}
\alias{plot_compare_joint}
\title{Plot and compare joint posterior distribution(s) from ABC-SMC-(D)RF result}
\usage{
plot_compare_joint(
  plots = NULL,
  abc_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  lims = NULL,
  nBins = 5
)
}
\arguments{
\item{plots}{An existed ABC-SMC-(D)RF joint plots result.
If provided, plot_compare_joint() will plot the new ABC-SMC-(D)RF result and compare it with provided plots results.
If plots = NULL, plot_compare_joint() will make a new plot for the ABC-SMC-(D)RF result.}

\item{abc_results}{An ABC-SMC-(D)RF result.
Will be plotted out by the function.}

\item{parameters_truth}{A dataframe containing true values of parameters from the ground-truth distributions.
If provided, the function will plot the true values or distributions of parameters.}

\item{parameters_labels}{A dataframe containing labels in the plots for corresponding parameters.
If provided, parameter labels will be exhibited on the plots' axes.}

\item{lims}{A dataframe containing the maximum and minimum bounds for parameters.
If provided, x-axis and y-axis will be scaled by them.}

\item{nBins}{Number of contour bins shown in the plot. Default is 5.}
}
\value{
An ABC-SMC-(D)RF joint plots object `plots` containing the joint plots results of posterior distributions.
The user can use the function to compare ABC-SMC-(D)RF joint plots with the joint posterior distribution(s)
of other ABC-SMC-(D)RF result(s).
}
\description{
`plot_compare_joint()` plots the joint posterior distribution(s) for the provided ABC-SMC-(D)RF result.
It can also compare the joint posterior distributions for the provided ABC-SMC-(D)RF result with several ABC-SMC-(D)RF plots results.
}
