% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_compare_qqplot}
\alias{plot_compare_qqplot}
\title{.......}
\usage{
plot_compare_qqplot(
  plots = NULL,
  abc_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  statistics_labels = NULL,
  sample_num = NULL,
  plot_statistics = FALSE,
  plot_hist = FALSE,
  lims = NULL,
  alpha = 0.3,
  plot_prior = FALSE
)
}
\description{
.......
}
