% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smcrf.r
\name{plot_smcrf_joint}
\alias{plot_smcrf_joint}
\title{`plot_smcrf_joint()` plots the joint distribution(s) of two parameters for each iteration from an
Approximate Bayesian Computation sequential Monte Carlo via random forest result.}
\usage{
plot_smcrf_joint(
  smcrf_results,
  parameters_truth = NULL,
  parameters_labels = NULL,
  lims = NULL,
  nBins = 5
)
}
\arguments{
\item{smcrf_results}{An ABC-SMC-(D)RF result containing the inference distributions of parameters from each iteration.}

\item{parameters_truth}{A dataframe containing true values or the ground-truth distributions of parameters.
If provided, the function will plot the true values or distributions of parameters.}

\item{parameters_labels}{A dataframe containing labels in the plots for corresponding parameters.
If provided, parameter labels will exhibit on the plots' axes.}

\item{lims}{A dataframe containing the maximum and minimum bounds for parameters.
If provided, x-axis and y-axis will be scaled by them.}

\item{nBins}{Number of contour bins shown in the plot. Default is 5.}
}
\description{
`plot_smcrf_joint()` plots the joint distribution(s) of two parameters for each iteration from an
Approximate Bayesian Computation sequential Monte Carlo via random forest result.
}
\examples{
lims <- data.frame(
    parameter = c("theta", "mu"),
    min = c(0, 0),
    max = c(10, 10)
)
}
