% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcrf.r
\name{smcrf}
\alias{smcrf}
\title{Approximate Bayesian Computation sequential Monte Carlo via random forests
for marginal or joint posterior distributions.}
\usage{
smcrf(
  method = "smcrf-single-param",
  statistics_target = NULL,
  smcrf_existed_results = NULL,
  model,
  perturb,
  range = NULL,
  parameters_initial = NULL,
  nParticles,
  parallel,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{method}{Random forest method to implement in each iteration ("smcrf-single-param" by default).
method = "smcrf-single-param": implements ABC-RF for each parameter and results in their marginal posterior distributions.
method = "smcrf-multi-param": implements ABC-DRF for all parameters and results in the joint posterior distribution.}

\item{statistics_target}{A dataframe containing statistics from data.
Column names are the statistics IDs.}

\item{smcrf_existed_results}{An existing ABC-SMC-RF result (NULL by default).
If provided, smcrf will continue ABC-SMC-RF from the last iteration of the previous run.}

\item{model}{Model for the statistics.
The function must take two inputs: parameters and parallel.}

\item{perturb}{.}

\item{range}{.}

\item{parameters_initial}{.}

\item{nParticles}{.}

\item{parallel}{.}

\item{n_cores}{.}

\item{...}{.}
}
\value{
Something.
}
\description{
Approximate Bayesian Computation sequential Monte Carlo via random forests
for marginal or joint posterior distributions.
}
